Integrate Global Confirm Modal Component into Alert Dialog Component
