--- apps/web/src/app/_components/global-confirm-modal/global-confirm-modal.css.ts
@@ -1,7 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-export const buttonWrapper = style({
-  display: 'flex',
-  justifyContent: 'flex-end',
-  gap: '8px',
-});
