--- apps/web/src/app/_components/global-confirm-modal/index.ts
@@ -1,1 +1,0 @@
-export * from './global-confirm-modal';
